updated lwmqtt

+++ src/lwmqtt/client.c
@@ -71,7 +71,7 @@ static lwmqtt_err_t lwmqtt_read_from_network(lwmqtt_client_t *client, size_t off
   // read while data is missing
   while (read < len) {
     // get remaining time
-    uint32_t remaining_time = client->timer_get(client->command_timer);
+    int32_t remaining_time = client->timer_get(client->command_timer);
 
     // check timeout
     if (remaining_time <= 0) {
@@ -81,7 +81,7 @@ static lwmqtt_err_t lwmqtt_read_from_network(lwmqtt_client_t *client, size_t off
     // read
     size_t partial_read = 0;
     lwmqtt_err_t err = client->network_read(client->network, client->read_buf + offset + read, len - read,
-                                            &partial_read, remaining_time);
+                                            &partial_read, (uint32_t)remaining_time);
     if (err != LWMQTT_SUCCESS) {
       return err;
     }
@@ -100,7 +100,7 @@ static lwmqtt_err_t lwmqtt_write_to_network(lwmqtt_client_t *client, size_t offs
   // write while data is left
   while (written < len) {
     // get remaining time
-    uint32_t remaining_time = client->timer_get(client->command_timer);
+    int32_t remaining_time = client->timer_get(client->command_timer);
 
     // check timeout
     if (remaining_time <= 0) {
@@ -110,7 +110,7 @@ static lwmqtt_err_t lwmqtt_write_to_network(lwmqtt_client_t *client, size_t offs
     // read
     size_t partial_write = 0;
     lwmqtt_err_t err = client->network_write(client->network, client->write_buf + offset + written, len - written,
-                                             &partial_write, remaining_time);
+                                             &partial_write, (uint32_t)remaining_time);
     if (err != LWMQTT_SUCCESS) {
       return err;
     }
